refactor: move some ExitId tests to PaymentExitGame test suite to expose tested functions through public API
